src/main/java/reladder/Point.java
package reladder;

import java.util.Objects;

public class Point {
    private int index;
    private Direction direction;

    public Point(int index, Direction direction) {
        this.index = index;
        this.direction = direction;
    }

    public int move() {
        if (this.direction.isLeft()) {
            return index - 1;
        }

        if (this.direction.isRight()) {
            return index + 1;
        }

        return this.index;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Point point = (Point) o;
        return index == point.index &&
                Objects.equals(direction, point.direction);
    }

    @Override
    public int hashCode() {
        return Objects.hash(index, direction);
    }
}
